+++ MoEngagePluginCards/DataTransfer/NativeToHybrid/MoEngageCardsData.swift
@@ -12,13 +12,15 @@ extension MoEngageCardsData: HybridEncodable {
         static let showAllTab = "shouldShowAllTab"
         static let cardCategories = "categories"
         static let cards = "cards"
+        static let accessibility = "accessibility"
     }
 
     func encodeForHybrid() -> [String: Any?] {
         return [
             HybridKeys.showAllTab: self.showAllTab,
             HybridKeys.cardCategories: self.cardCategories,
-            HybridKeys.cards: self.cards.encodeForHybrid()
+            HybridKeys.cards: self.cards.encodeForHybrid(),
+            HybridKeys.accessibility: self.staticImageAccessibilityData?.mapValues { $0.encodeForHybrid() } as Any
         ]
     }
 }